requirements.txt
annotated-types==0.7.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
arrow==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
flexcache==0.3
flexparser==0.4
httplib2==0.31.2
idna==3.11
oauth-flow==1.0.3
oauth2==1.9.0.post1
Pint==0.25.2
platformdirs==4.9.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlalchemy==2.0.36
stravalib==2.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
